feat(pyapi): restructure python API
