+++ pyapi/pyproject.toml
@@ -5,3 +5,6 @@ requires-python = ">=3.7"
 [build-system]
 requires = ["maturin>=1.0,<2.0"]
 build-backend = "maturin"
+
+[tool.maturin]
+python-source = "python"
