+++ pyapi/python/rustsat/__init__.pyi
