+++ pyapi/python/rustsat/encodings/card/__init__.py
@@ -1,0 +1,5 @@
+from .rustsat.encodings.card import *
+
+__doc__ = rustsat.encodings.card.__doc__
+if hasattr(rustsat.encodings.card, "__all__"):
+    __all__ = rustsat.encodings.card.__all__
