+++ pyapi/python/rustsat/encodings/card/__init__.pyi
@@ -1,0 +1,12 @@
+from rustsat import Lit, Cnf, VarManager
+from typing import List, final
+
+@final
+class Totalizer:
+    def __new__(cls, lits: List[Lit]) -> Totalizer: ...
+    def extend(self, lits: List[Lit]) -> None: ...
+    def n_lits(self) -> int: ...
+    def n_clauses(self) -> int: ...
+    def n_vars(self) -> int: ...
+    def encode_ub(self, min_ub: int, max_ub: int, var_manager: VarManager) -> Cnf: ...
+    def enforce_ub(self, ub) -> List[Lit]: ...
