+++ pyapi/python/rustsat/encodings/pb/__init__.py
@@ -1,0 +1,5 @@
+from .rustsat.encodings.pb import *
+
+__doc__ = rustsat.encodings.pb.__doc__
+if hasattr(rustsat.encodings.pb, "__all__"):
+    __all__ = rustsat.encodings.pb.__all__
