+++ pyapi/python/rustsat/encodings/pb/__init__.pyi
@@ -2,16 +2,6 @@ from rustsat import Lit, Cnf, VarManager
 from typing import List, final
 
 @final
-class Totalizer:
-    def __new__(cls, lits: List[Lit]) -> Totalizer: ...
-    def extend(self, lits: List[Lit]) -> None: ...
-    def n_lits(self) -> int: ...
-    def n_clauses(self) -> int: ...
-    def n_vars(self) -> int: ...
-    def encode_ub(self, min_ub: int, max_ub: int, var_manager: VarManager) -> Cnf: ...
-    def enforce_ub(self, ub) -> List[Lit]: ...
-
-@final
 class GeneralizedTotalizer:
     def __new__(cls, wlits: List[tuple[int, Lit]]) -> GeneralizedTotalizer: ...
     def extend(self, lits: List[Lit]) -> None: ...
@@ -38,8 +28,6 @@ class BinaryAdder:
     def n_vars(self) -> int: ...
     def encode_ub(self, min_ub: int, max_ub: int, var_manager: VarManager) -> Cnf: ...
     def enforce_ub(self, ub) -> List[Lit]: ...
-    def encode_lb(self, min_lb: int, max_lb: int, var_manager: VarManager) -> Cnf: ...
-    def enforce_lb(self, lb) -> List[Lit]: ...
     def encode_lb(self, min_lb: int, max_lb: int, var_manager: VarManager) -> Cnf: ...
     def enforce_lb(self, lb) -> List[Lit]: ...
     def encode_both(
